refactor: Rename request package to http
