+++ src/mg/emberframework/annotation/http/Get.java
@@ -1,4 +1,4 @@
-package mg.emberframework.annotation.request;
+package mg.emberframework.annotation.http;
 
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
