+++ src/mg/emberframework/annotation/http/Url.java
@@ -1,4 +1,4 @@
-package mg.emberframework.annotation.request;
+package mg.emberframework.annotation.http;
 
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -7,5 +7,6 @@ import java.lang.annotation.Retention;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Post {
+public @interface Url {
+    String value();
 }
