+++ src/mg/emberframework/core/data/RequestVerb.java
@@ -13,7 +13,7 @@ public class RequestVerb {
 
     public static String getMethodVerb(Method method) {
         String verb = RequestVerb.GET;
-        if (method.isAnnotationPresent(mg.emberframework.annotation.request.Post.class)) {
+        if (method.isAnnotationPresent(mg.emberframework.annotation.http.Post.class)) {
             verb = RequestVerb.POST;
         }
         return verb;
